Fix the macOS build, and the macOS/Linux bugs the audit found behind it
CI's macOS job caught what no local build could: macos_upstream_smoke.rs is
`#![cfg(target_os = "macos")]`, so on Windows it compiles to nothing and its
SetupOptions literal quietly kept the shape it had before aedfcb8 added
`reaper_language_package`. That one field is the actual build fix.

Since this is the third time code the Windows toolchain never compiles has
gone stale, I audited the whole non-Windows surface rather than just the
file CI happened to stop at. Two real macOS/Linux defects fell out, both
shipped in 0.4.0:

The DataView tree's set_value never refreshed the controls below the package
list. Its Windows twin funnels every toggle route through
refresh_after_packages_toggle, whose tail syncs the OSARA keymap note, the
Spanish variant picker and the REAPER-language dropdown; the non-Windows
path only synced them from on_selection_changed, which does not fire when
you toggle the row that already has focus. So arrowing to a row and pressing
Space left all three stale — and because the language dropdown owns its
*contents* rather than just its enabled state, that would have shipped the
new language-pack picker permanently empty on macOS, silently installing
whichever pack came first. The widgets are created after the tree (creation
order is tab order, so they cannot move above it), so they now reach the
model through a cell, the same way the self-referential model handle
already does.

And the Configuration group checkbox counted already-applied steps as
unticked members, so "all ticked" was never true once any step was applied
— the normal case, an existing ReaPack remote counts. The group drew as
unchecked forever and, since the toggle renderer inverts the displayed
value, could only ever check everything and never uncheck. Windows'
compute_configuration_group_tristate had the `!already_applied` filter all
along; the DataView twin never got it.

Also fixes the four warnings the macOS runner has been printing, and a
duplicated sync call I left in aedfcb8.

Worth recording: `cargo clippy -p rabbit-platform --target aarch64-apple-darwin`
does work on Windows and reproduces the macOS warnings exactly. rabbit-core
cannot cross-check — ring and libsqlite3-sys need a macOS C compiler — and
the UI crate needs the wx toolchain, so CI stays the only check for those.

+++ crates/rabbit-platform/src/signature.rs
@@ -213,7 +213,7 @@ mod tests {
     #[cfg(unix)]
     fn exit_status(code: c_int) -> ExitStatus {
         use std::os::unix::process::ExitStatusExt;
-        ExitStatus::from_raw((code as i32) << 8)
+        ExitStatus::from_raw(code << 8)
     }
 
     #[cfg(windows)]
